Fix Facebook leaking internal URIs
Facebook is now leaking internal URIs into the data download. Instead
of correct local references into the zip archive(s), URIs to the
non-publicly resolvable "interncache-frc.fbcdn.net" are present in the
zip files. Including what looks like tokens or signatures.

The fix for this is atrociously painful. Perhaps one day Facebook will
actually have a test suite for this monstrosity.

This adds about 2 to 3 seconds to start-up time on my computer with my
most recent data download. It's closer to 10 seconds when not-running
the fix in parallel.